pyproject.toml
[tool.poetry]
name = "mandelbrot-renderer"
version = "0.1.0"
description = ""
authors = ["Bhavin <bhavin200@hotmail.co.uk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
numpy = "^2.2.1"
matplotlib = "^3.10.0"
pillow = "^11.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
